python_quiz/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.7
peewee==3.13.3
pyTelegramBotAPI==3.7.2
requests==2.20.0
six==1.15.0
urllib3==1.24.3
pyyaml==5.3.1
